requirements.txt
kagglehub
pandas
seaborn
matplotlib
numpy
langchain-community
langchain-text-splitters
langchain-openai
langchain-chroma
python-dotenv
torch
transformers
tqdm
gradio
